requirements.txt
selenium==4.4.3
pytest
pytest-xdist
allure-pytest
webdriver-manager
